refactor: extract classes FileDescriptorSetLoader and EventTypeGenerator
